prototype/requirements.txt
flask
pdfplumber
beautifulsoup4
sentence-transformers
pymilvus
python-dotenv
tika
scikit-learn
flask-cors
faiss-cpu